refactor: DRY up byline tests and remove confusing Array in snapshots
