+++ packages/article-byline/__tests__/android/article-byline.android.test.js
@@ -1,5 +1,6 @@
+import ArticleByline from "../../src/article-byline";
 import shared from "../shared";
 
 describe("ArticleByline tests on android", () => {
-  shared();
+  shared(ArticleByline);
 });
