+++ packages/article-byline/__tests__/ios/article-byline.ios.test.js
@@ -1,5 +1,6 @@
+import ArticleByline from "../../src/article-byline";
 import shared from "../shared";
 
 describe("ArticleByline tests on iOS", () => {
-  shared();
+  shared(ArticleByline);
 });
