+++ packages/article-byline/__tests__/web/article-byline-with-links.web.test.js
@@ -1,5 +1,6 @@
-import shared from "../shared-with-links";
+import { ArticleBylineWithLinks } from "../../src/article-byline";
+import shared from "../shared";
 
 describe("ArticleBylineWithLinks tests on web", () => {
-  shared();
+  shared(ArticleBylineWithLinks);
 });
