+++ packages/article-byline/__tests__/web/article-byline.web.test.js
@@ -1,5 +1,6 @@
+import ArticleByline from "../../src/article-byline";
 import shared from "../shared";
 
 describe("ArticleByline tests on web", () => {
-  shared();
+  shared(ArticleByline);
 });
